+++ playground/examples/PluginExample.tsx
@@ -1,17 +1,10 @@
 import * as THREE from "three"
 import type { Meta } from "types.ts"
-import {
-  createPlugin,
-  createT,
-  EventPlugin,
-  Resource,
-  useFrame,
-  useThree,
-} from "../../src/index.ts"
+import { createT, EventPlugin, plugin, Resource, useFrame, useThree } from "../../src/index.ts"
 import { OrbitControls } from "../controls/OrbitControls.tsx"
 
 // LookAt plugin - works for all Object3D elements
-const LookAtPlugin = createPlugin().prop(THREE.Object3D, element => ({
+const LookAtPlugin = plugin().prop(THREE.Object3D, element => ({
   lookAt: (target: THREE.Object3D | [number, number, number]) => {
     useFrame(() => {
       if (Array.isArray(target)) {
@@ -24,7 +17,7 @@ const LookAtPlugin = createPlugin().prop(THREE.Object3D, element => ({
 }))
 
 // Shake plugin - works for both Camera and Light elements using array syntax
-const ShakePlugin = createPlugin().prop([THREE.Camera, THREE.DirectionalLight], element => ({
+const ShakePlugin = plugin().prop([THREE.Camera, THREE.DirectionalLight], element => ({
   shake: (intensity = 0.1) => {
     const originalPosition = element.position.clone()
     useFrame(() => {
@@ -36,7 +29,7 @@ const ShakePlugin = createPlugin().prop([THREE.Camera, THREE.DirectionalLight], 
 }))
 
 // Custom filter plugin - works for objects with a 'material' property using type guard
-const MaterialPlugin = createPlugin().prop(
+const MaterialPlugin = plugin().prop(
   (element: any): element is THREE.Mesh =>
     element instanceof THREE.Mesh && element.material !== undefined,
   element => ({
@@ -52,7 +45,7 @@ const MaterialPlugin = createPlugin().prop(
 )
 
 // Global plugin - applies to all elements using single argument
-const GlobalPlugin = createPlugin().prop((element, context) => ({
+const GlobalPlugin = plugin().prop(element => ({
   log: (message: string) => {
     console.log(`[${element.constructor.name}] ${message}`)
   },
