+++ src/index.ts
@@ -4,8 +4,8 @@ export { $S3C } from "./constants.ts"
 export { createEntity, createT } from "./create-t.tsx"
 export { EventPlugin } from "./event-plugin.ts"
 export { useFrame, useThree } from "./hooks.ts"
+export { plugin } from "./plugin.ts"
 export { useProps } from "./props.ts"
 export * from "./raycasters.tsx"
 export * as S3 from "./types.ts"
-export { createPlugin } from "./types.ts"
 export { autodispose, getMeta, hasMeta as hasMeta, load, meta } from "./utils.ts"
