+++ src/types.ts
@@ -291,54 +291,7 @@ export interface Plugin<TFn = (element: any) => any> {
   (): TFn
 }
 
-/**
- * Creates a plugin with a unified prop API
- * Usage:
- * - createPlugin(() => { setup }).prop((element, context) => methods) // apply to all elements
- * - createPlugin(() => { setup }).prop(Constructor, (element, context) => methods) // single constructor filter
- * - createPlugin(() => { setup }).prop([Constructor1, Constructor2], (element, context) => methods) // multiple constructors
- * - createPlugin(() => { setup }).prop((element): element is T => condition, (element, context) => methods) // type guard
- */
-// Helper function to create the actual plugin implementation
-function createFilteredPlugin(
-  setup: (() => any) | undefined,
-  filterArg: any,
-  methods: any,
-): Plugin<any> {
-  const plugin: Plugin<any> = () => {
-    // Run setup once if provided and store result as context
-    const context = setup ? setup() : undefined
-
-    return ((element: any) => {
-      // Handle single constructor
-      if (typeof filterArg === "function" && filterArg.prototype) {
-        if (element instanceof filterArg) {
-          return methods(element, context)
-        }
-      }
-      // Handle array of constructors
-      else if (Array.isArray(filterArg)) {
-        for (const Constructor of filterArg) {
-          if (element instanceof Constructor) {
-            return methods(element, context)
-          }
-        }
-      }
-      // Handle type guard function
-      else if (typeof filterArg === "function") {
-        if (filterArg(element)) {
-          return methods(element, context)
-        }
-      }
-
-      return {}
-    }) as any
-  }
-
-  return plugin
-}
-
-interface PluginBuilder<TContext extends object> {
+export interface PluginBuilder<TContext extends object> {
   // Apply to all elements - single argument
   prop<Methods extends Record<string, any>>(
     methods: (element: any, context: TContext) => Methods,
@@ -373,34 +326,6 @@ interface PluginBuilder<TContext extends object> {
     (element: T): Methods
     (element: any): {}
   }>
-}
-
-export function createPlugin<TContext extends object>(
-  setup?: () => TContext,
-): PluginBuilder<TContext> {
-  return {
-    // Unified prop method - handles both single and two argument cases
-    prop(filterArgOrMethods: any, methods?: any): Plugin<any> {
-      // Single argument case - apply to all elements
-      if (methods === undefined) {
-        type PluginFn = (element: any) => any
-
-        const plugin: Plugin<PluginFn> = () => {
-          // Run setup once if provided and store result as context
-          const context = setup ? setup() : undefined
-
-          return ((element: any) => {
-            return filterArgOrMethods(element, context)
-          }) as PluginFn
-        }
-
-        return plugin
-      }
-
-      // Two argument case - use filtering
-      return createFilteredPlugin(setup, filterArgOrMethods, methods)
-    },
-  } as PluginBuilder<TContext>
 }
 
 export type InferPluginProps<T, TPlugins extends Plugin[]> = Merge<{
@@ -531,7 +456,6 @@ export type Props<T, TPlugins extends Plugin[] | undefined = Plugin[]> = Partial
   Merge<
     [
       MapToRepresentation<InstanceOf<T>>,
-      EventHandlers,
       {
         args: T extends Constructor ? ConstructorOverloadParameters<T> : undefined
         attach: string | ((parent: object, self: Meta<InstanceOf<T>>) => () => void)
